+++ app/models/tasks/mandate_hold_task.rb
@@ -14,6 +14,8 @@
 # CAVC Remands Overview: https://github.com/department-of-veterans-affairs/caseflow/wiki/CAVC-Remands
 
 class MandateHoldTask < Task
+  include CavcTimedHoldConcern
+
   VALID_PARENT_TYPES = [
     CavcTask
   ].freeze
@@ -24,16 +26,7 @@ class MandateHoldTask < Task
 
   def self.create_with_hold(parent_task)
     ActiveRecord::Base.transaction do
-      mandate_hold_task = create!(parent: parent_task, appeal: parent_task.appeal)
-      mandate_hold_task.create_timed_hold_task
-      mandate_hold_task
-    end
-  end
-
-  def update_timed_hold
-    ActiveRecord::Base.transaction do
-      children.open.where(type: :TimedHoldTask).last&.cancelled!
-      create_timed_hold_task
+      create!(parent: parent_task, appeal: parent_task.appeal).tap { |mhold_task| mhold_task.create_timed_hold_task }
     end
   end
 
@@ -53,27 +46,9 @@ class MandateHoldTask < Task
     TASK_ACTIONS
   end
 
-  def create_timed_hold_task
-    days_to_hold = decision_date_plus_90_days
-    if days_to_hold > 0
-      TimedHoldTask.create_from_parent(
-        self,
-        days_on_hold: days_to_hold,
-        instructions: [COPY::MANDATE_HOLD_TASK_DEFAULT_INSTRUCTIONS]
-      )
-    end
-  end
-
   private
 
   def set_assignee
     self.assigned_to = CavcLitigationSupport.singleton if assigned_to.nil?
-  end
-
-  def decision_date_plus_90_days
-    decision_date = appeal.cavc_remand.decision_date
-    end_date = decision_date + 90.days
-    # convert to the number of days from today
-    (end_date - Time.zone.today).to_i
   end
 end
